pyproject.toml
[project]
name = "staffeli-nt"
version = "0.3.1"
description = "Staffeli NT Technology"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [
    { name = "Ken Friis Larsen" },
    { name = "Mads Obitsø" },
    { name = "Mathias Hall-Andersen" },
    { name = "William Due" },
    { name = "Jacob Herbst" }
]

dependencies = [
    "canvasapi>=3.4.0",
    "rich>=14.2.0",
    "ruamel-yaml>=0.16.10",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0",
    "pyright>=1.1.407",
    "types-requests",
    "types-PyYAML",
    "ruff>=0.8.0",
]

[project.scripts]
staffeli = "staffeli_nt.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["staffeli_nt"]

[tool.mypy]
incremental = false
check_untyped_defs = true
disallow_untyped_defs = false
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle errors, pyflakes, isort

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
